fix(mria_rlog_server): avoid protocol version checking log spam
Fixes #26 .

Currently, the protocol version check is done directly in
`mria_rlog_server`, but the last check is not cached. If a server in
the cluster has a different protocol version, the logs will
periodically be spammed with warning messages about version
discrepancies.

By moving this check into `mria_lb`, making it remember the last
version checked and warning only when the version changes, we can
minimize log noise.