fix(update_counter): don't call `update_counter` directly
Due to possible race conditions when bootstrapping a replicant, it
might be possible to reach an inconsistent state.

Minimal scenario to trigger the race condition: imagine you have a
table with a single counter C that initially has value of 0. It gets
incremented once. Bootstrap can go in two ways: 0) start of bootstrap
1) full scan reads value of C equals 0. replica is initialized with
value 0 2) counter gets incremented 3) replicant receives the TLOG op
and updates its value of C to 1. All good.  But it can also go like
this: 0) start of bootstrap 1) counter gets incremented 2) full scan
reads value of C equal 1, replica is initialized with value 1 3)
replicant receives the TLOG op and updates the value to 2.

Thanks to @ieQu1 / @k32 for catching this. 🍻